feat: GlobalErrorBoundary 구현
