+++ src/App.tsx
@@ -1,12 +1,9 @@
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { Suspense, useEffect } from 'react';
-import { ErrorBoundary } from 'react-error-boundary';
+import { useEffect } from 'react';
 import { Toaster } from 'react-hot-toast';
 import { RouterProvider } from 'react-router-dom';
-import Error from '@/pages/error/Error';
 import { router } from '@/routes/router.tsx';
-import Loading from './pages/loading/Loading';
 import queryClient from './queryClient';
 import './shared/styles/index.css';
 
@@ -34,11 +31,9 @@ const App = () => {
   return (
     <QueryClientProvider client={queryClient}>
       <ReactQueryDevtools initialIsOpen={false} />
-      <ErrorBoundary FallbackComponent={Error}>
-        <Suspense fallback={<Loading />}>
-          <RouterProvider router={router} />
-        </Suspense>
-      </ErrorBoundary>
+
+      <RouterProvider router={router} />
+
       <Toaster containerStyle={{ margin: '0 auto' }} />
     </QueryClientProvider>
   );
