+++ src/layout/Layout.tsx
@@ -1,5 +1,6 @@
 import { Outlet, ScrollRestoration, useLocation } from 'react-router-dom';
 import { ROUTES_CONFIG } from '@/routes/routesConfig';
+import GlobalErrorBoundary from '@/shared/components/GlobalErrorBoundary/GlobalErrorBoundary';
 import Header from '@/shared/components/Header/Header';
 
 const Layout = () => {
@@ -11,11 +12,11 @@ const Layout = () => {
   const shouldShowHeader = !isSearchPath && !isOnboardingPath;
 
   return (
-    <>
+    <GlobalErrorBoundary>
       {shouldShowHeader && <Header />}
       <Outlet />
       <ScrollRestoration />
-    </>
+    </GlobalErrorBoundary>
   );
 };
